Thu, Sep 22, 2022, 1:45 AM +00:00

+++ pythontut/operators.py
@@ -1,4 +1,4 @@
-weight = float(input("Weight: "))
+weight = int(input("Weight: "))
 unit = input("(K)g or (L)b: ")
 
 if unit.lower == "L":
